Fade in/out elements when added to the DOM
Fade in is easy to achieve with pure CSS, fade out - not without JavaScript. I've used the elm-style-animation to, so far, fade out the PAR Help infopanel.

I haven't extended the fadeout effect to other elements on the page though, such as the SD line (the part that shows the selector for PAR), or fading between the different images of Prince (which I'm not sure if it's possible).